renamed files: qrCode -> claimQrCode, scanQrCode -> scanClaimQrCode
